fix for requested changes: - make cache -> _cache - mv ws and http functions to their own files
