+++ api/networks/ethereum/index.js
@@ -5,13 +5,18 @@ import { EventEmitter } from 'events'
 
 /*
   should do caching
+
+  TODO'S:
+  - http polling subscriptions
+  - filterers
 */
 
 export default class EthereumNetworkProvider extends Provider {
   constructor ({ endpoint }) {
     super({ endpoint })
-    this.cache = {}
     this.subscriptions = {}
+    // private vars
+    this._cache = {}
     this._subscriptionIds = {}
   }
 
@@ -22,14 +27,13 @@ export default class EthereumNetworkProvider extends Provider {
       const [ name ] = params
       const subscription = this.subscriptions[name]
       if (subscription) return subscription.id
-      if (!this.socket) await this.connect()
 
       const id = await super.request(request)
       // map for subcription id's
       this._subscriptionIds[id] = name
       // acctual subcription interface
       this.subscriptions[name] = new Subscription({ name, id })
-      this.socket.addEventListener('message', this.subscriptions[name].handler.bind(this.subscriptions[name]))
+      this.transport.socket.addEventListener('message', this.subscriptions[name].handler.bind(this.subscriptions[name]))
       return id
     } else if (request.method === 'eth_unsubscribe') {
       const params = request.params
@@ -40,7 +44,7 @@ export default class EthereumNetworkProvider extends Provider {
       // removes listener and if their are no subcriptions left unsubcribes from the node
       if (handler) subscription.removeListener('update', handler)
       if (!subscription.listeners('update').length) {
-        this.socket.removeEventListener('message', subscription.handler.bind(subscription))
+        this.transport.socket.removeEventListener('message', subscription.handler.bind(subscription))
         const confirmation = await super.request({ method: 'eth_unsubscribe', params: [subscription.id] })
         return confirmation
       }
@@ -52,34 +56,32 @@ export default class EthereumNetworkProvider extends Provider {
       if (!params || !params.length) key = request.method
       else key = [request.method, ...params].join('/')
 
-      if(!this.cache[key]) {
-        this.cache[key] = await super.request(request)
+      if(!this._cache[key]) {
+        this._cache[key] = await super.request(request)
       }
 
-      return this.cache[key]
+      return this._cache[key]
     }
   }
 
   async connect () {
-    const socket = await super.connect()
     // subcscribe to new heads for caching purposes
     await this.request({
       method: 'eth_subscribe',
       params: ['newHeads'],
     })
     this.subscriptions.newHeads.on('update', (state) => {
-      this.cache = {
+      this._cache = {
         eth_blockNumber: state.number
       }
     })
-    return socket
   }
 
   async close () {
     Object.values(this.subscriptions).forEach((subscription) => {
       try {
         subscription.removeAllListners('update')
-        this.socket.removeEventListener('message',subscription.handler.bind(subscription))
+        this.transport.socket.removeEventListener('message',subscription.handler.bind(subscription))
         this.request({ method: 'eth_unsubscribe', params: [subscription.id] })
       } catch (e) {
         console.error(e)
